removal of unused variables/typedef (#7)
* include/sycl/algorithm/
	- sort.hpp
		replace int by size_t in lines 87, 88
	- exclusive_scan.hpp
		replace int by size_t in lines 114, 115
	- inner_product.hpp
		remove lines 128, 129 : unused variables (localid, globalid)
	- tranform_reduce.hpp
		remove line 66 : unused type
		remove lines 85, 86 : unused variables (localid, globalid)
	- inclusive_scan.hpp
		replace int by size_t in lines 92, 93
	- count_if.hpp
		remove line 66 : unused type
		remove lines 86, 87 : unused variables (localid, globalid)
	- for_each.hpp
		remove line 53 : unused type

* find.hpp : remove unused variables

algorithm/find.hpp remove unused variables lines 139, 140 (localid, globalid)

+++ include/sycl/algorithm/exclusive_scan.hpp
@@ -111,8 +111,8 @@ OutputIterator exclusive_scan(ExecutionPolicy &sep, InputIterator b,
       h.parallel_for<
           cl::sycl::helpers::NameGen<1, typename ExecutionPolicy::kernelName> >(
           r, [aI, aO, bop, vectorSize, i](cl::sycl::nd_item<3> id) {
-            int td = 1 << (i - 1);
-            int m_id = id.get_global(0);
+            size_t td = 1 << (i - 1);
+            size_t m_id = id.get_global(0);
             if (m_id < vectorSize && m_id >= td) {
               aO[m_id] = bop(aI[m_id - td], aI[m_id]);
             } else {
