+++ include/sycl/algorithm/inclusive_scan.hpp
@@ -89,8 +89,8 @@ OutputIterator inclusive_scan(ExecutionPolicy &sep, InputIterator b,
           outBuf->template get_access<cl::sycl::access::mode::read_write>(h);
       h.parallel_for<typename ExecutionPolicy::kernelName>(
           r, [aI, aO, bop, vectorSize, i](cl::sycl::nd_item<3> id) {
-            int td = 1 << (i - 1);
-            int m_id = id.get_global(0);
+            size_t td = 1 << (i - 1);
+            size_t m_id = id.get_global(0);
 
             if (m_id < vectorSize && m_id >= td) {
               aO[m_id] = bop(aI[m_id - td], aI[m_id]);
